LayoutTests/editing/execCommand/list-wrapping-image-crash-expected.txt

CONSOLE MESSAGE: line 9: Error: HIERARCHY_REQUEST_ERR: DOM Exception 3
